# backend/solarhub-nestjs/swagger-output.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/shop/b2b/products": {
      "get": {
        "operationId": "B2bCatalogController_getB2bProducts",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "بحث نصي في اسم المنتج أو الوصف",
            "schema": {
              "example": "panel",
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب التصنيف",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "company_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب الشركة البائعة",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "min_price",
            "required": false,
            "in": "query",
            "description": "أقل سعر",
            "schema": {
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "max_price",
            "required": false,
            "in": "query",
            "description": "أعلى سعر",
            "schema": {
              "example": 5000,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "رقم الصفحة",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page_size",
            "required": false,
            "in": "query",
            "description": "عدد العناصر في كل صفحة",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "B2B Catalog"
        ]
      }
    },
    "/api/v1/shop/b2b/search": {
      "get": {
        "operationId": "B2bCatalogController_searchB2bProducts",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "بحث نصي في اسم المنتج أو الوصف",
            "schema": {
              "example": "panel",
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب التصنيف",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "company_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب الشركة البائعة",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "min_price",
            "required": false,
            "in": "query",
            "description": "أقل سعر",
            "schema": {
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "max_price",
            "required": false,
            "in": "query",
            "description": "أعلى سعر",
            "schema": {
              "example": 5000,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "رقم الصفحة",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page_size",
            "required": false,
            "in": "query",
            "description": "عدد العناصر في كل صفحة",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "B2B Catalog"
        ]
      }
    },
    "/api/v1/shop/b2b/companies/{id}/products": {
      "get": {
        "operationId": "B2bCatalogController_getCompanyProducts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "بحث نصي في اسم المنتج أو الوصف",
            "schema": {
              "example": "panel",
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب التصنيف",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "company_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب الشركة البائعة",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "min_price",
            "required": false,
            "in": "query",
            "description": "أقل سعر",
            "schema": {
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "max_price",
            "required": false,
            "in": "query",
            "description": "أعلى سعر",
            "schema": {
              "example": 5000,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "رقم الصفحة",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page_size",
            "required": false,
            "in": "query",
            "description": "عدد العناصر في كل صفحة",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "B2B Catalog"
        ]
      }
    },
    "/api/v1/shop/b2b/categories/{type}/{id}/products": {
      "get": {
        "operationId": "B2bCatalogController_getCategoryProducts",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "بحث نصي في اسم المنتج أو الوصف",
            "schema": {
              "example": "panel",
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب التصنيف",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "company_id",
            "required": false,
            "in": "query",
            "description": "فلتر حسب الشركة البائعة",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "min_price",
            "required": false,
            "in": "query",
            "description": "أقل سعر",
            "schema": {
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "max_price",
            "required": false,
            "in": "query",
            "description": "أعلى سعر",
            "schema": {
              "example": 5000,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "رقم الصفحة",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page_size",
            "required": false,
            "in": "query",
            "description": "عدد العناصر في كل صفحة",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "B2B Catalog"
        ]
      }
    },
...
    "/api/v1/shop/company-customers": {
      "get": {
        "operationId": "CompanyCustomersController_findAll",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Customers (CRM)"
        ]
      },
      "post": {
        "operationId": "CompanyCustomersController_create",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerFormDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Customers (CRM)"
        ]
      }
    },
    "/api/v1/shop/company-customers/{id}": {
      "get": {
        "operationId": "CompanyCustomersController_findOne",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Customers (CRM)"
        ]
      },
      "put": {
        "operationId": "CompanyCustomersController_update",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerFormDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Customers (CRM)"
        ]
      },
      "delete": {
        "operationId": "CompanyCustomersController_remove",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Customers (CRM)"
        ]
      }
    },
...
  },
  "info": {
    "title": "SolarHub API",
    "description": "The SolarHub Backend API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "basic": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "CustomerFormDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}